Moved loadFiles and loadFilesErr action from visStateUpdater to composite actions and reducers
