+++ src/actions/vis-state-actions.js
@@ -420,34 +420,6 @@ export function setFilterPlot(idx, newProp) {
 }
 
 /**
- * Trigger file loading dispatch `addDataToMap` if succeed, or `loadFilesErr` if failed
- * @memberof visStateActions
- * @param {Array<Object>} files array of fileblob
- * @returns {{type: ActionTypes.LOAD_FILES, files: *}}
- * @public
- */
-export function loadFiles(files) {
-  return {
-    type: ActionTypes.LOAD_FILES,
-    files
-  };
-}
-
-/**
- * Trigger loading file error
- * @memberof visStateActions
- * @param {*} error
- * @returns {{type: ActionTypes.LOAD_FILES_ERR, error: *}}
- * @public
- */
-export function loadFilesErr(error) {
-  return {
-    type: ActionTypes.LOAD_FILES_ERR,
-    error
-  };
-}
-
-/**
  * This declaration is needed to group actions in docs
  */
 /**
