+++ src/reducers/combined-updaters.js
@@ -20,11 +20,15 @@
 
 import {fitBoundsUpdater} from './map-state-updaters';
 import {toggleModalUpdater} from './ui-state-updaters';
-import {updateVisDataUpdater as visStateUpdateVisDataUpdater} from './vis-state-updaters';
+import {loadFilesErrUpdater as loadFilesErrVisUpdater, updateVisDataUpdater as visStateUpdateVisDataUpdater} from './vis-state-updaters';
 import {receiveMapConfigUpdater as stateMapConfigUpdater} from './map-state-updaters';
 import {receiveMapConfigUpdater as styleMapConfigUpdater} from './map-style-updaters';
 import {findMapBounds} from 'utils/data-utils';
 import KeplerGlSchema from 'schemas';
+import Task, {withTask} from 'react-palm/tasks';
+import {addDataToMap, loadFilesErr} from 'actions';
+import {processFileToLoad} from 'utils/file-utils';
+import {LOAD_FILE_TASK} from 'tasks/tasks';
 
 // compose action to apply result multiple reducers, with the output of one
 
@@ -182,3 +186,63 @@ export const addDataToMapUpdater = (state, {payload}) => {
 };
 
 export const addDataToMapComposed = addDataToMapUpdater;
+
+/**
+ * Trigger file loading dispatch `addDataToMap` if succeed, or `loadFilesErr` if failed
+ * @memberof combinedUpdaters
+ * @param {Object} state kepler.gl instance state, containing all subreducer state
+ * @param {Object} action action
+ * @param {Array<Object>} action.files array of fileblob
+ * @returns {Object} nextState
+ * @public
+ */
+export const loadFilesUpdater = (state, action) => {
+  const {files} = action;
+
+  const filesToLoad = files.map(fileBlob => processFileToLoad(fileBlob));
+
+  // reader -> parser -> augment -> receiveVisData
+  const loadFileTasks = [
+    Task.all(filesToLoad.map(LOAD_FILE_TASK)).bimap(
+      results => {
+        const data = results.reduce((f, c) => ({
+          // using concat here because the current datasets could be an array or a single item
+          datasets: f.datasets.concat(c.datasets),
+          // we need to deep merge this thing unless we find a better solution
+          // this case will only happen if we allow to load multiple keplergl json files
+          config: {
+            ...f.config,
+            ...(c.config || {})
+          }
+        }), {datasets: [], config: {}, options: {centerMap: true}});
+        return addDataToMap(data);
+      },
+      error => loadFilesErr(error)
+    )
+  ];
+
+  return withTask(
+    {
+      ...state,
+      visState: {
+        ...state.visState,
+        fileLoading: true
+      }
+    },
+    loadFileTasks
+  );
+};
+
+/**
+ * Trigger loading file error
+ * @memberof combinedUpdaters
+ * @param {Object} state `visState`
+ * @param {Object} action action
+ * @param {*} action.error
+ * @returns {Object} nextState
+ * @public
+ */
+export const loadFilesErrUpdater = (state, action) => ({
+  ...state,
+  visState: loadFilesErrVisUpdater(state.visState, action)
+});
