+++ src/reducers/composers.js
@@ -27,7 +27,9 @@ import * as combinedUpdaters from './combined-updaters';
  */
 const actionHandler = {
   [ActionTypes.UPDATE_VIS_DATA]: combinedUpdaters.updateVisDataUpdater,
-  [ActionTypes.ADD_DATA_TO_MAP]: combinedUpdaters.addDataToMapUpdater
+  [ActionTypes.ADD_DATA_TO_MAP]: combinedUpdaters.addDataToMapUpdater,
+  [ActionTypes.LOAD_FILES]: combinedUpdaters.loadFilesUpdater,
+  [ActionTypes.LOAD_FILES_ERR]: combinedUpdaters.loadFilesErrUpdater
 };
 
 export default actionHandler;
