+++ src/reducers/vis-state-updaters.js
@@ -1291,51 +1291,9 @@ function closeSpecificMapAtIndex(state, action) {
   };
 }
 
-/**
- * Trigger file loading dispatch `addDataToMap` if succeed, or `loadFilesErr` if failed
- * @memberof visStateUpdaters
- * @param {Object} state `visState`
- * @param {Object} action action
- * @param {Array<Object>} action.files array of fileblob
- * @returns {Object} nextState
- * @public
- */
-export const loadFilesUpdater = (state, action) => {
-  const {files} = action;
-
-  const filesToLoad = files.map(fileBlob => processFileToLoad(fileBlob));
-
-  // reader -> parser -> augment -> receiveVisData
-  const loadFileTasks = [
-    Task.all(filesToLoad.map(LOAD_FILE_TASK)).bimap(
-      results => {
-        const data = results.reduce((f, c) => ({
-          // using concat here because the current datasets could be an array or a single item
-          datasets: f.datasets.concat(c.datasets),
-          // we need to deep merge this thing unless we find a better solution
-          // this case will only happen if we allow to load multiple keplergl json files
-          config: {
-            ...f.config,
-            ...(c.config || {})
-          }
-        }), {datasets: [], config: {}, options: {centerMap: true}});
-        return addDataToMap(data);
-      },
-      error => loadFilesErr(error)
-    )
-  ];
-
-  return withTask(
-    {
-      ...state,
-      fileLoading: true
-    },
-    loadFileTasks
-  );
-};
 
 /**
- * Trigger loading file error
+ * Update state with file loading error
  * @memberof visStateUpdaters
  * @param {Object} state `visState`
  * @param {Object} action action
